ISingleTextboxPageModel wip
