+++ src/FamilyHubs.Referral.Web/Pages/ProfessionalReferral/Email.cshtml
@@ -1,17 +1,19 @@
 ﻿@page
 @model FamilyHubs.Referral.Web.Pages.ProfessionalReferral.EmailModel
 @{
-    ViewData["Title"] = Model.PartialTextBoxViewModel.HeadingText;
+    ViewData["Title"] = Model.HeadingText;
 }
 
 @section Back {
     <a asp-page="/ProfessionalReferral/ContactDetails" asp-route-serviceId="@Model.ServiceId" asp-route-serviceName="@Model.ServiceName" class="govuk-back-link">Back</a>
 }
 
+@*todo: move rest into partial? except maybe not form?? and back?? and title??*@
+@*or can we specify the whole cshtml to use directly, and not use a partial?*@
 <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
         <form method="post">
-            <partial name="../Shared/_TextBoxDetail" for="PartialTextBoxViewModel">
+            <partial name="_SingleTextboxPage" for="@Model" />
         </form>
     </div>
 </div>
