+++ src/FamilyHubs.Referral.Web/Pages/ProfessionalReferral/SupportDetails.cshtml
@@ -11,7 +11,7 @@
 <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
         <form method="post">
-            <partial name="_TextBoxDetail" for="@Model" />
+            <partial name="_SingleTextboxPage" for="@Model" />
         </form>
     </div>
 </div>
