+++ src/FamilyHubs.Referral.Web/Pages/ProfessionalReferral/SupportDetails.cshtml.cs
@@ -1,25 +1,11 @@
 using FamilyHubs.Referral.Core.DistributedCache;
 using FamilyHubs.Referral.Core.Helper;
 using FamilyHubs.Referral.Core.Models;
+using FamilyHubs.Referral.Web.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace FamilyHubs.Referral.Web.Pages.ProfessionalReferral;
-
-public interface ISingleTextboxPageModel
-{
-    //todo: get rid of text and label
-    public string HeadingText { get; set; }
-    //todo: no need for this?
-    public string ErrorId { get; set; }
-    public string HintText { get; set; }
-    public string TextBoxLabel { get; set; }
-    public string MainErrorText { get; set; }
-    //todo: default to mainerror
-    public string? TextBoxErrorText { get; set; }
-    public string? TextBoxValue { get; set; }
-    public bool ValidationValid { get; set; }
-}
 
 public class SupportDetailsModel : PageModel, ISingleTextboxPageModel
 {
@@ -27,12 +13,9 @@ public class SupportDetailsModel : PageModel, ISingleTextboxPageModel
     public string? ServiceId { get; private set; }
     public string? ServiceName { get; private set; }
 
-    //todo: separate static with changing
-    //public PartialTextBoxViewModel PartialTextBoxViewModel { get; } = new()
-    //{
     public string ErrorId { get; set; } = "error-summary-title";
     public string HeadingText { get; set; } = "Who should the service contact in the family?";
-    public string HintText { get; set; } = "This must be a person aged 16 or over.";
+    public string? HintText { get; set; } = "This must be a person aged 16 or over.";
     public string TextBoxLabel { get; set; } = "Full name";
     public string MainErrorText { get; set; } = "Enter a full name";
     public string? TextBoxErrorText { get; set; } = "Enter a full name";
@@ -59,20 +42,15 @@ public class SupportDetailsModel : PageModel, ISingleTextboxPageModel
 
         if (!string.IsNullOrEmpty(model?.FamilyContactFullName))
         {
-            //todo: two?
-            //PartialTextBoxViewModel.TextBoxValue = model.FamilyContactFullName;
             TextBoxValue = model.FamilyContactFullName;
         }
     }
 
     public async Task<IActionResult> OnPostAsync(string serviceId, string serviceName)
     {
-        if (!ModelState.IsValid)
+        if (!ModelState.IsValid || string.IsNullOrWhiteSpace(TextBoxValue))
         {
-            //PartialTextBoxViewModel.TextBoxValue = TextBoxValue;
-            if (string.IsNullOrWhiteSpace(TextBoxValue))
-                ValidationValid = false;
-
+            ValidationValid = false;
             return Page();
         }
