+++ src/FamilyHubs.Referral.Web/Pages/Shared/_SingleTextboxPage.cshtml
@@ -1,4 +1,4 @@
-﻿@model FamilyHubs.Referral.Web.Pages.ProfessionalReferral.ISingleTextboxPageModel
+﻿@model FamilyHubs.Referral.Web.Models.ISingleTextboxPageModel
 <!-- This control can only be used when there are no other inputs on the page -->
 @if (!Model.ValidationValid)
 {
